refs #243 - remove extra executable target for errors and move CustomNSError tests to ExitCodeTests
