+++ Tests/ArgumentParserUnitTests/ExitCodeTests.swift
@@ -76,3 +76,53 @@ extension ExitCodeTests {
     }
   }
 }
+
+// MARK: - CustomNSError tests
+
+extension ExitCodeTests {
+  enum MyCustomNSError: CustomNSError {
+    case myFirstCase
+    case mySecondCase
+
+    var errorCode: Int {
+      switch self {
+      case .myFirstCase:
+        return 101
+      case .mySecondCase:
+        return 102
+      }
+    }
+
+    var errorUserInfo: [String : Any] {
+      switch self {
+      case .myFirstCase:
+        return [NSLocalizedDescriptionKey: "My first case localized description"]
+      case .mySecondCase:
+        return [:]
+      }
+    }
+  }
+
+  struct CheckFirstCustomNSErrorCommand: ParsableCommand {
+
+    @Option
+    var errorCase: Int
+
+    func run() throws {
+      switch errorCase {
+      case 101:
+        throw MyCustomNSError.myFirstCase
+      default:
+        throw MyCustomNSError.mySecondCase
+      }
+    }
+  }
+  
+  func testCustomErrorCodeForTheFirstCase() {
+    XCTAssertEqual(CheckFirstCustomNSErrorCommand.exitCode(for: MyCustomNSError.myFirstCase), ExitCode(rawValue: 101))
+  }
+
+  func testCustomErrorCodeForTheSecondCase() {
+    XCTAssertEqual(CheckFirstCustomNSErrorCommand.exitCode(for: MyCustomNSError.mySecondCase), ExitCode(rawValue: 102))
+  }
+}
